updated database schema, added more templates

+++ src/main/java/com/menulabs/taxisplus/repositories/UnidadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.menulabs.taxisplus.domain.Unidad;
 
-public interface UnidadRepository extends JpaRepository<Unidad, Integer> {
+public interface UnidadRepository extends JpaRepository<Unidad, Long> {
 
 }
